% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euraxess.R
\name{read_job_offer}
\alias{read_job_offer}
\title{read_job_offer}
\usage{
read_job_offer(url)
}
\arguments{
\item{url}{: Url of a job offer.}
}
\value{
datatable with the content of the job offer.
}
\description{
Read the job offer passed as url from euraxess and extract :
- the general text on it
- the content of the tickle box
- two parameter of location

The text is extracted in a broad way using "#node-" + id of the job offer, it is the only way to generalize to
all the jobs offers.
}
\examples{
\dontrun{
url <- "https://euraxess.ec.europa.eu/jobs/498418"
df<-read_job_offer(url)
}
}
