% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euraxess.R
\name{resume_job_offer}
\alias{resume_job_offer}
\title{resume_job_offer}
\usage{
resume_job_offer(df, tagger, key_words)
}
\arguments{
\item{df}{: df of job offer scrapped by read_offer_offer.}

\item{tagger}{: a udpipe model for english annotation.}

\item{key_words}{: key words to create the summary with pagerank}
}
\value{
a summary of the job offer that can be empty of no sentence contain the keywords.
}
\description{
Summarized a job offer using textrank. During the execution of the function, the terminology is
reduced to the keywords passed as input to get a summary containing only sentence with keywords of interrest.
}
\examples{
\dontrun{
key_words<-c("systematic", "review", "text", "mining", "text")
tagger <- udpipe_load_model("english-gum-ud-2.4-190531.udpipe")
resume_job_offer(df, tagger, key_words)


}


}
