% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euraxess.R
\name{scrape_urls_euraxess}
\alias{scrape_urls_euraxess}
\title{scrape_urls_euraxess}
\usage{
scrape_urls_euraxess(search_url, last_page)
}
\arguments{
\item{search_url}{: url of the first page of results of a job offers search.}

\item{last_page}{: last page of the search to be scrapped using this function. 3 mean that the first page of resuls
will be scrapped, then the three others.}
}
\value{

}
\description{
Function recollect all the url of job offers from euraxes to run read_job_offer.
}
\examples{
\dontrun{
scrape_urls_euraxess(search_url, 3)
urls<-scrape_urls_euraxess(search_url, 3)
job_offer<-pblapply(urls, read_job_offer)
}

}
