% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euraxess.R
\name{scrape_urls_page}
\alias{scrape_urls_page}
\title{scrape_urls_page}
\usage{
scrape_urls_page(search_url)
}
\arguments{
\item{search_url}{:  Url of one results page of a search for job offer. Can be the first or page 17. Passed as
input by scrape scrape_urls_euraxess.}
}
\value{

}
\description{
Function to scrape all the urls of job from result page. Allow to used some filter from euraxess,
like "early stage researcher". Used in scrape_urls_euraxess.
}
\examples{
\dontrun{
search_url <- "https://euraxess.ec.europa.eu/jobs/search/field_research_profile/first-stage-researcher-r1-446?sort=created&order=desc"
urls <- scrape_urls_pages(search_url)
}
}
