% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euraxess.R
\name{summarise_all_job_offers}
\alias{summarise_all_job_offers}
\title{summarise_all_job_offers}
\usage{
summarise_all_job_offers(job_offer, tagger, key_words)
}
\arguments{
\item{job_offer}{: data.table of job offer.}

\item{tagger}{: an updpipe model for the annotation}

\item{key_words}{: key words on which construct the summary}
}
\value{

}
\description{
Run resume_job_offer on all the job_offers and return a data.table with the summary
}
\examples{
\dontrun{
key_words<-c("systematic", "review", "text", "mining", "text")
tagger <- udpipe_load_model("english-gum-ud-2.4-190531.udpipe")
urls<-scrape_urls_euraxess(search_url, 3)
job_offer<-pblapply(urls, read_job_offer)
job_offer<-rbindlist(job_offer, use.names = TRUE)
saveRDS(job_offer, file="job_offer.RDS")
job_offer<-summarise_all_job_offers(job_offer, tagger, key_words)
saveRDS(job_offer, file="job_offer_curated.RDS")
}
}
